actualizacion permisos

+++ proyecto/src/main/java/lbj/king/proyecto/api_rest/PlayRestController.java
@@ -5,7 +5,9 @@ import java.sql.SQLException;
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,8 +16,10 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import lbj.king.proyecto.DTO.PlayDTO;
+import lbj.king.proyecto.model.Play;
 import lbj.king.proyecto.services.PlayService;
 import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;
 
@@ -54,7 +58,16 @@ public class PlayRestController {
 	}
 
     @DeleteMapping("/{id}")
-    public PlayDTO deletePlay(@PathVariable long id) {
-        return playService.deletePlay(id);
+    public ResponseEntity<?> deletePlay(@PathVariable Long id) {
+    String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+    Play play = playService.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+    if (!play.getUser().getName().equals(username)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN);
     }
+
+    playService.deletePlay(id);
+    return ResponseEntity.noContent().build();
+}
 }
